+++ HMI_PanelSaw/App.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using HMI_PanelSaw.Views;
+using System;
 using System.Collections.Generic;
 using System.Configuration;
 using System.Data;
@@ -13,5 +14,12 @@ namespace HMI_PanelSaw
     /// </summary>
     public partial class App : Application
     {
+        protected override void OnStartup(StartupEventArgs e)
+        {
+            base.OnStartup(e);
+            LoginView loginWindow = new LoginView();
+            loginWindow.Show();
+        }
     }
+
 }