+++ HMI_PanelSaw/Views/ParametersView.xaml.cs
@@ -21,6 +21,7 @@ namespace HMI_PanelSaw.Views
     public partial class ParametersView : UserControl
     {
         private AdsService _adsService;
+        private AuthService _authService;
         private const float MIN_PANEL_DIMENSION = 10f;     
         private const float MAX_PANEL_DIMENSION = 5000f;    
         private const float MIN_BLADE_DIAMETER = 100f;      
@@ -28,11 +29,13 @@ namespace HMI_PanelSaw.Views
         private const float MIN_BLADE_THICKNESS = 1f;       
         private const float MAX_BLADE_THICKNESS = 10f;      
 
-        public ParametersView(AdsService adsService)
+        public ParametersView(AdsService adsService, AuthService authService)
         {
             InitializeComponent();
             _adsService = adsService;
+            _authService = authService;
             InitializeParameterHandles();
+            LoadUserParameters();
             LoadParametersFromPLC();
         }
         private void InitializeParameterHandles()
@@ -54,6 +57,7 @@ namespace HMI_PanelSaw.Views
                 MessageBox.Show($"Error initializing parameters {ex.Message}","Initialization Error", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
+
         private void LoadParametersFromPLC()
         {
             try
@@ -73,6 +77,16 @@ namespace HMI_PanelSaw.Views
                 MessageBox.Show($"Error loading parameters from PLC {ex.Message}", "Loading error", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
+
+        private void LoadUserParameters()
+        {
+            txtUserId.Text = $"Id: {_authService.CurrentUser.Id}";
+            txtUserUsername.Text = $"Username: {_authService.CurrentUser.Username}";
+            txtUserRole.Text = $"Role : {_authService.CurrentUser.Role}";
+            txtUserCreatedAt.Text = $"Created At: {_authService.CurrentUser.CreatedAt}";
+            txtUserLastLogin.Text = $"Last Login At: {_authService.CurrentUser.LastLoginAt}";
+        }
+
         private bool TryParseAndValidate(string input, string parameterName, float min, float max, out float result)
         {
             result = 0f;
@@ -156,5 +170,6 @@ namespace HMI_PanelSaw.Views
                 SaveParametersToPLC();
             }
         }
+    
     }
 }